+++ cli/src/fmt/simple/mod.rs
@@ -12,6 +12,7 @@ mod account;
 mod asset;
 mod effect;
 mod ledger;
+mod offer;
 mod orderbook;
 mod payment;
 mod trade_aggregation;