+++ cli/src/main.rs
@@ -97,6 +97,18 @@ fn build_app<'a, 'b>() -> App<'a, 'b> {
                             )
                     ),
 
+                )
+                .subcommand(
+                    listable!(
+                        SubCommand::with_name("offers")
+                            .about("Fetch all open offers for an account")
+                            .arg(
+                                Arg::with_name("ID")
+                                    .required(true)
+                                    .help("The identifier of the account to look up"),
+                            )
+                    ),
+
                 ),
         )
         .subcommand(
@@ -394,6 +406,7 @@ fn main() {
             ("details", Some(sub_m)) => account::details(&client, sub_m),
             ("transactions", Some(sub_m)) => account::transactions(&client, sub_m),
             ("effects", Some(sub_m)) => account::effects(&client, sub_m),
+            ("offers", Some(sub_m)) => account::offers(&client, sub_m),
             _ => return print_help_and_exit(),
         },
         ("assets", Some(sub_m)) => match sub_m.subcommand() {